+++ src/app/shared/shared-modals/components/sign-message/sign-message.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SignMessageComponent } from './sign-message.component';
+
+describe('SignMessageComponent', () => {
+  let component: SignMessageComponent;
+  let fixture: ComponentFixture<SignMessageComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [SignMessageComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(SignMessageComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});