+++ src/environments/environment.extension.ts
@@ -25,7 +25,7 @@ export const environment = {
     networkPassphrase: Networks.TESTNET,
     canRemove: false,
   }],
-  version: '1.33.1',
+  version: '1.40.0',
 };
 
 /*