+++ src/environments/environment.prod.ts
@@ -26,5 +26,5 @@ export const environment = {
     networkPassphrase: Networks.PUBLIC,
     canRemove: true,
   }],
-  version: '1.33.1'
+  version: '1.40.0'
 };