+++ src/extension/background/background.ts
@@ -4,7 +4,7 @@ import {
   RuntimeMessage,
   RuntimeResponse,
   XBULL_CONNECT_BACKGROUND, XBULL_GET_NETWORK_BACKGROUND,
-  XBULL_GET_PUBLIC_KEY_BACKGROUND,
+  XBULL_GET_PUBLIC_KEY_BACKGROUND, XBULL_SIGN_MESSAGE_BACKGROUND,
   XBULL_SIGN_XDR_BACKGROUND,
 } from '../interfaces';
 import { requestConnection } from '~extension/background/connection.background';
@@ -12,6 +12,7 @@ import { requestPublicKey } from '~extension/background/public-key.background';
 import { requestSignXDR } from '~extension/background/sign-transaction.background';
 import { getWindowId, setWindowId } from '~extension/background/state.background';
 import { requestNetwork } from '~extension/background/get-network.background';
+import { requestSignMessage } from '~extension/background/sign-message.background';
 
 (browser?.action || browser?.browserAction).onClicked.addListener(async () => {
   let windowId: number | undefined;
@@ -95,6 +96,10 @@ browser.runtime.onMessage.addListener(async (message: RuntimeMessage, sender): P
 
     case XBULL_GET_NETWORK_BACKGROUND:
       return requestNetwork()
+        .catch(catchError);
+
+    case XBULL_SIGN_MESSAGE_BACKGROUND:
+      return requestSignMessage(message)
         .catch(catchError);
 
     default: