+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.40.0] - 2025-03-30
+### Add
+- Add logic to sign messages following SEP-0053
+
+### Change
+- Upgrade the Stellar SDK to the latest version (compatible with protocol 23)
+
 ## [1.33.1] - 2025-03-30
 ### Change
 - Upgrade the SorobanDomains SDK library to 0.2.0